// meepo/components/coach.service.d.ts
import { ApiService } from '../meepo-fox';
import { Observable } from 'rxjs/Observable';
export declare class CoachService {
    api: ApiService;
    roles: string[];
    coach: any;
    widget: any;
    form: any;
    day: number;
    year: number;
    month: number;
    hasSelect: any[];
    timeList: any[];
    showTiaokuan: boolean;
    hasInit: boolean;
    showJoin: boolean;
    showCitys: boolean;
    tags: any;
    city: any;
    joinForm: any;
    footer: any[];
    header: any;
    constructor(api: ApiService);
    selectTabs(e: any): void;
    getCity(): Observable<any>;
    getCitys(): Observable<any>;
    getHotCitys(): Observable<any>;
    setCity(res: any): void;
    onInit(cache?: boolean): void;
    init2(): void;
    onSelectTag(e: any): void;
    onTextareaChange(): void;
    init(): void;
    gethasSelect(): void;
    daySelect(e: any): void;
    onSelect(e: any): void;
    post(): void;
    getNowFormatDate(date: any, fmt: string): string;
    updateCoach(): Observable<any>;
    checkHasRole(role: string, roles: string[]): boolean;
    getSkillGroup(): Observable<any>;
    skillJoin(data: any): Observable<any>;
    getSkillList(): Observable<any>;
}